libtinyiiod: fix mask calculation
An "OPEN" command sent by iio-oscilloscope can look like this:
OPEN cf-ad9361-dds-core-lpc\0008190 0000000f CYCLIC\0
where:
device is: "cf-ad9361-dds-core-lpc"
number of samples is: "0008190"
mask for 4 channels is: "0000000f"

After the mask, in our case 0000000f, we also have the CYCLIC option,
so the return value of the function will be "-EINVAL" after this test:
if (str[8] != '\0')
	return -EINVAL;

When extracting the mask
else if (c >= 'a' && c <= 'f')
	c -= 'a' - 0x10;
is executed, resulting c = 0x15, which is wrong, it should be 0xf.
So the simplest fix would be replacing the implementation, since is wrong, with:
mask = strtoul(str, NULL, 16);
since strtol() is already used in the libtinyiiod.

Signed-off-by: Cristian Pop <cristian.pop@analog.com>