libtinyiiod: Use “stdint” types
Necessary so that when we use it with ad9361, cppcheck won't complain,
otherwise casts between types need to be used.
Add "compat.h" that will implement stdint types and
a smaller version of stdlib.
This is meant to be used with smaller uC, when the compiler
doesn’t come with “stdlib”.
Further implementation will be necessary, to be completely usable.
Functions like “malloc”, “free”, “snprintf” need to be implemented.
A cmake option was added, to select the smaller version of stdlib.

Signed-off-by: Cristian Pop <cristian.pop@analog.com>