tinyiiod: change return types for {write,read} callbacks
This change breaks API a bit, but in many cases this should cause compiler
warnings. The read & write callbacks should return how much was
read/written, or an error if it occurs.

This should help the library be a little more robust, with the intent of
propagating error-codes, in such a way that they can be better interpreted,
by doing retries or failing completely.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>